zmiany w api
