+++ flask-server/database_stuff/app.py
@@ -1,28 +1,22 @@
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
+
 db = SQLAlchemy()
 
 
-
 def create_app():
-
     app = Flask(__name__, template_folder='templates')
     app.config['SQLALCHEMY_DATABASE_URI'] = "mysql+pymysql://root:password@localhost/housedb"
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = True
 
-
     db.init_app(app)
-    from views import views
-    from auth import auth
-    
-    app.register_blueprint(views, url_prefix='/')
-    app.register_blueprint(auth, url_prefix='/')
-    
+
     migrate = Migrate(app, db)
     app.app_context().push()
     return app
-    
+
+
 '''
 WAŻNE!!!
 Żeby coś robić musicie mieć u siebie folder migrations
@@ -30,7 +24,7 @@ WAŻNE!!!
     flask db init
 Cały folder jest w gitignore i niech tak zostanie
 
-'''    
+'''
 
 '''
 Żeby wymigrować zmiany w modelach do bazy to robicie tak
